# docs/openapi.yaml
openapi: 3.1.0
info:
  title: SatL Engine API
  description: |-
    The Docker Engine REST API v1.43 as `satld` implements it -- the only external surface SatL has (invariant #8). Every intentional divergence from Docker carries a numbered entry in `docs/api-compat.md`, cited inline throughout this document.

    The daemon listens on the unix socket `/var/run/satl.sock`. There is no TCP listener for this API: the cluster's own protocol (`satl.internal.v1`) is a separate, node-to-node-only surface on ports 2377 and 2378.

    ## Version prefixes

    Every operation is reachable both on its bare path (`/info`) and under a Docker version prefix (`/v1.43/info`). Prefixes are accepted from 1.24 up to 1.43; a prefix outside that range is a `400` in Docker's error shape, not a 404. A leading segment that merely starts with `v` but is not `v<major>.<minor>` -- `/v1/info` -- is not version negotiation at all and routes (and 404s) as an ordinary path, where Docker would parse `1` as a version.

    ## Responses this document cannot attach to an operation

    - **Unknown path**: `404` with the body `{"message": "page not found"}`. Every Docker endpoint SatL does not implement answers this shape -- attach, commit, export, rename, restart, pause, update, top, stats, changes, archive, build, push, and the rest of the `/images/{name}/*` family (api-compat #22, #144, #152).
...
    - **Out-of-range version prefix**: `400` with Docker's error body, minted before routing.

...

...

    ## Error shape
...
servers:
...
paths:
...
  /secrets:
    get:
      tags:
      - Secret
      summary: '`GET /secrets`.'
      description: 'A secret document never carries its payload: the renderer has no way to emit one, so no handler here can leak one by forgetting a flag (invariant #7).'
      operationId: SecretList
      parameters:
      - name: filters
        in: query
        description: 'A non-empty value is rejected with 501 rather than silently listing everything (api-compat #47).'
        required: false
        schema:
          type: string
      responses:
        '200':
          description: One row per secret, payloads excluded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SecretResponse'
        '500':
          description: Daemon error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '501':
          description: '`?filters=` was non-empty, or the daemon has no store wired.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '503':
          description: This node is not a swarm manager, or no manager is reachable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
  /secrets/create:
    post:
      tags:
      - Secret
      summary: '`POST /secrets/create`.'
      description: 'The payload arrives base64-encoded and is never written to a worker''s disk: it is delivered over mTLS and materialized on a per-task tmpfs that dies with the jail (invariant #7). This is one of the two endpoints whose body *is* a payload, so it accepts a larger request body than the rest of the API.'
      operationId: SecretCreate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SecretSpecWire'
        required: true
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdResponse'
        '400':
          description: Invalid name, payload or body size.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '409':
          description: A secret of that name already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: Daemon error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '501':
          description: Not implemented by this daemon.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '503':
          description: This node is not a swarm manager, or no manager is reachable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
  /secrets/{id}:
    get:
      tags:
      - Secret
      summary: '`GET /secrets/{id}`.'
      description: 'One secret, payload excluded (invariant #7).'
      operationId: SecretInspect
      parameters:
      - name: id
        in: path
        description: Secret ID or name.
        required: true
        schema:
          type: string
      responses:
        '200':
          description: The secret document, payload excluded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecretResponse'
        '404':
          description: No such secret.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: Daemon error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '501':
          description: Not implemented by this daemon.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '503':
          description: This node is not a swarm manager, or no manager is reachable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
    delete:
      tags:
      - Secret
      summary: '`DELETE /secrets/{id}`.'
      description: Removes a secret. A secret still referenced by a service is refused.
      operationId: SecretDelete
      parameters:
      - name: id
        in: path
        description: Secret ID or name.
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Removed.
        '404':
          description: No such secret.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '409':
          description: The secret is still referenced by a service.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: Daemon error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '501':
          description: Not implemented by this daemon.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '503':
          description: This node is not a swarm manager, or no manager is reachable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
  /secrets/{id}/update:
    post:
      tags:
      - Secret
      summary: '`POST /secrets/{id}/update`: refused.'
      description: '**Always 501.** Docker''s endpoint can only change labels -- the payload is immutable there too -- and answering "updated" to a request that changed nothing an operator cares about is worse than refusing it. Rotation is the documented path: create a new secret, update the services that use it, remove the old one.'
      operationId: SecretUpdate
      parameters:
      - name: id
        in: path
        description: Secret ID or name.
        required: true
        schema:
          type: string
      responses:
        '501':
          description: Secrets are immutable; rotate instead.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
...
components:
  schemas:
...
    ErrorBody:
      type: object
      description: 'Docker error envelope: every non-2xx response body is `{"message": "..."}`.'
      required:
      - message
      properties:
        message:
          type: string
          description: Human-readable error message.
...
    IdResponse:
      type: object
      description: |-
        Docker's `IDResponse`: the answer to `POST /secrets/create` and `POST
        /configs/create`. Spelled `ID`, like [`ServiceCreateResponse`].
      properties:
        ID:
          type: string
          description: The new object's ID.
          default: ''
...
    ObjectVersionWire:
      type: object
      description: |-
        Object version envelope (`{"Index": 42}`), Docker's optimistic-concurrency
        token.
      properties:
        Index:
          type: integer
          format: int64
          description: Store version of the object.
          default: 0
          minimum: 0
...
    SecretResponse:
      type: object
      description: Docker's `Secret` document (`GET /secrets`, `GET /secrets/{id}`).
      properties:
        CreatedAt:
          type: string
          description: Creation time, RFC 3339 nanoseconds.
          default: ''
        ID:
          type: string
          description: Secret ID.
          default: ''
        Spec:
          oneOf:
          - $ref: '#/components/schemas/SecretSpecWire'
            description: The spec, always without its payload.
          default:
            Name: ''
        UpdatedAt:
          type: string
          description: Last update time, RFC 3339 nanoseconds.
          default: ''
        Version:
          oneOf:
          - $ref: '#/components/schemas/ObjectVersionWire'
            description: Store version.
          default:
            Index: 0
    SecretSpecWire:
      type: object
      description: |-
        `SecretSpec`: the document `POST /secrets/create` takes and every secret
        response echoes back — minus `Data`, which is never rendered.
      properties:
        Data:
          type:
          - string
          - 'null'
          description: Base64-encoded payload. Required on create, never rendered.
          default: null
        Driver:
          description: External secret driver — rejected (no driver plugins).
        Labels:
          type: object
          description: Free-form labels.
          default: {}
          additionalProperties:
            type: string
          propertyNames:
            type: string
        Name:
          type: string
          description: Secret name (unique across the cluster).
          default: ''
        Templating:
          description: Templating driver — rejected (no template engine).
...